Converted data cleaning script to cloud function, created infra for function and data
