+++ cloud_functions/cleaning/business_license_data/requirements.txt
@@ -1,0 +1,6 @@
+numpy
+pandas
+shapely
+db-dtypes
+google-cloud-bigquery
+google-cloud-bigquery-storage
