+++ infra/iam.tf
@@ -47,6 +47,15 @@ resource "google_cloudfunctions_function_iam_member" "business_licenses_invoker"
   member = "allUsers"
 }
 
+resource "google_cloudfunctions_function_iam_member" "clean_business_licenses_invoker" {
+  project        = google_cloudfunctions_function.clean_business_licenses_function.project
+  region         = google_cloudfunctions_function.clean_business_licenses_function.region
+  cloud_function = google_cloudfunctions_function.clean_business_licenses_function.name
+
+  role   = "roles/cloudfunctions.invoker"
+  member = "allUsers"
+}
+
 ################# FOOD INSPECTION DATA
 
 resource "google_cloudfunctions_function_iam_member" "food_inspections_invoker" {
